+++ client/src/components/Layout/index.tsx
@@ -1,15 +1,15 @@
 import { Outlet } from "react-router-dom";
 import { PhoneNumberProvider } from "@/contexts/phoneNumberContext";
-import { ScrollAnimationProvider } from "@/contexts/scrollAnimationContext.tsx";
+import { ScrollHeaderStyleProvider } from "@/contexts/scrollHeaderStyleContext.tsx";
 import Header from "../Header";
 
 export default function Layout() {
     return (
-        <ScrollAnimationProvider>
+        <ScrollHeaderStyleProvider>
             <PhoneNumberProvider>
                 <Header type="light" />
                 <Outlet />
             </PhoneNumberProvider>
-        </ScrollAnimationProvider>
+        </ScrollHeaderStyleProvider>
     );
 }
