+++ client/src/features/Main/LearnMore.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from "react";
 import CTAButton from "@/components/CTAButton";
-import { SectionKey } from "@/types/scrollAnimation.ts";
+import { SectionKey } from "@/types/scrollHeaderStyle.ts";
 
 interface LearnMoreProps {
     sectionId: SectionKey;
